update: move button generation to util and create new react msg
